Add --save-baseline and --baseline
Previously, Criterion.rs would store previous results in a single
directory, `.../base`. On every benchmark run, the `base`
directory would be overwritten. This behavior prevented comparison
to a known point while iterating on optimizations.

This change introduces the idea of baselines; a named directory
for results generated at some point in the past.  The default
behavior of using and overwriting the `base` directory is still
the same as before. However, these two new options allow the
user to explicitly name and compare to baselines.

`--save-baseline` creates and stores the results for that
benchmark run under a directory with the same name as that
baseline.

`--baseline` compares the results for the current run against
the baseline results. It loudly fails if the baseline is not
present for a specific benchmark.

Previous behavior is emulated using `--save-baseline base` when
no explicit baseline is named.

As discussed on #153
Closes #150 as a more flexible solution.