Display throughput changes in report
It can be tedious to determine how throughput changes based on timing
changes. The two are obviously directly related, but while it's obvious
a 50% reduction in time per iteration results in a 100% improvement in
throughput it's much more cumbersome to say a 26.3% improvement in
time results in a 35.6% improvement in throughput.

Sample output of the report now looks like:

bench_small_input/decode_slice/50
                        time:   [29.442 ns 29.456 ns 29.472 ns]
                        thrpt:  [1.5800 GiB/s 1.5809 GiB/s 1.5816 GiB/s]
                 change:
                        time:   [-17.930% -17.868% -17.814%] (p = 0.00 < 0.05)
                        thrpt:  [+21.676% +21.755% +21.847%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild