Use common cache of directories across runs
When running multiple benchmarks, we would keep a cache of
all_directories in Benchmark::run and check for duplicate names so they
could be de-duplicated.

In the case described in 272, run is called multiple time, which means
that the cache is empty on the second call.

Make this cache a member of the Criterion struct so it can be reused
across multiple benchmarks.

Fixes 272.