Fix trailing spaces in filenames on Windows
See the comment - basically the problem, is that Windows doesn't allow trailing spaces in filename, and instead silently trims them.

This works more or less well for file access, since the filename is trimmed in the same way both on creating and on access, but when used with directories, trying to create something like `dir /file` will simply fail with no apparent reason.

This was a problem before, but it was additionally revealed by the introduction of MAX_DIRECTORY_NAME_LEN since now name can be trimmed to a one with a trailing space much more frequently.