Add example of benchmarking code in a library crate.
Most of the examples include the function to be benchmarked in the benchmark file itself, but in practice that's not how benchmarking is done. This can be confusing to
developers new to rust, who don't know that benchmarks are treated as a separate crate, so it's helpful to document that.

Fixes #301.