Add option to load data instead of sampling it
Adds a new application argument:

--load-baseline <name>

This restores raw data from the implementation defined raw dumps and
subsequently treats it as if it had just been sampled. The dumps are
created for all runs early after the sampling and analysis process.
Note that this requires all benchmarks that are analyzed to have the
named baseline previously saved with --save-baseline (except the last
run which can be loaded under the name 'new').

The expose command line option requires a comparison baseline to be
specified as the option otherwise offers little advantage over the
already available csv exporter.